Lecture 1
Basic Arithmetic: Addition, Subtraction, Multiplication, Division. Different solutions for floating numbers division in python 2.
Powers, roots, order of operations, use parenthesis to specify orders.
Variable Assignments: ex. a = 5. Can reassignment. Use A to redefine A. 
Use object names to keep better track of what's going on in your code!

We learned some of the basics of numbers in Python. We also learned how to do arithmetic and use Python as a basic calculator. We then wrapped it up with learning about Variable Assignment in Python.